requirements.txt
dnspython
pyfiglet
dmarc
dkimpy
